Move log functions to utils.log
